+++ src/utils/earthengine.js
@@ -11,7 +11,6 @@ const workerOptions = [
     'data',
     'datasetId',
     'filter',
-    'format',
     'legend',
     'mask',
     'methods',