+++ src/utils/style.js
@@ -13,12 +13,15 @@ export const hoverStrokeMultiplier = 3
 export const eventStrokeColor = '#333333'
 export const clusterCountColor = '#000000'
 
-export const mapStyle = {
+export const defaultGlyphs =
+    'https://fonts.openmaptiles.org/{fontstack}/{range}.pbf'
+
+export const mapStyle = ({ glyphs = defaultGlyphs }) => ({
     version: 8,
     sources: {},
     layers: [],
-    glyphs: 'https://fonts.openmaptiles.org/{fontstack}/{range}.pbf', // TODO: Host ourseleves
-}
+    glyphs,
+})
 
 export default {
     textFont,